+++ lightorchestrator/service/vibe/ifaces/shifter.go
@@ -7,5 +7,5 @@ type Shifter interface {
 	Stabalizable
 
 	// Shift returns a value representing some change or shift based on t
-	Shift(t time.Time) float32
+	Shift(t time.Time) float64
 }
